breez-bench/kotlin: split Main.kt into Bench/Server/Funding/Lightning
Pure refactor — no behavioral changes. Main.kt is now ~150 lines of CLI
entry + arg parsing + smokeTest/traceSync + dispatch; each mode lives
next to the symbols it owns. Verified by a 10rps × 20s send_ln smoke
sweep (200/200 settled, 100%).
